worked on WeatherMappers.kt

+++ app/src/main/java/com/plcoding/weatherapp/domain/weather/WeatherType.kt
@@ -117,7 +117,7 @@ sealed class WeatherType(
     )
 
     companion object {
-        fun fromWMO(code: Int): WeatherType {
+        fun fromWMO(code: Double): WeatherType {
             return when(code) {
                 0 -> ClearSky
                 1 -> MainlyClear
